analyse/pyproject.toml
[tool.poetry]
name = "adersim-app-2-analyse"
version = "0.1.0"
description = "Data preprocessing and topic modeling for relevant YouTube video transcripts"
authors = ["Sharuka Thirimanne"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
deepmultilingualpunctuation = "1.0.1"
sentence-transformers = "3.2.1"
bertopic = "0.16.4"
matplotlib = "3.9.2"
google-api-python-client = "2.151.0"
dspy = "2.5.29"
google-auth = "2.36.0"
ipykernel = "6.29.5"
ipywidgets = "8.1.5"
youtube-transcript-api = "0.6.2"
langchain-ollama = "0.2.0"
langchain-qdrant = "0.2.0"
langchain-community = "0.3.5"
langdetect = "^1.0.9"
onnxruntime = "1.19.2"
qdrant-client = "^1.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
